Issue #43 refers: Removed duplicate classes -bzrsupport now depends on reahl-component and reahl-dev
