+++ reahl-bzrsupport/.reahlproject
@@ -3,11 +3,17 @@
     <package name="reahl"/>
   </namespaces>
 
+  <deps purpose="run">
+    <egg name="reahl-component"/>
+    <egg name="reahl-dev"/>
+  </deps>
+
   <deps purpose="test">
     <thirdpartyegg name="nose"/>
   </deps>
 
   <export entrypoint="setuptools.file_finders" name="reahl_finder" locator="reahl.bzrsupport:find_files"/> 
+  <export entrypoint="reahl.dev.xmlclasses" name="BzrSourceControl" locator="reahl.bzrsupport:BzrSourceControl"/>
 
   <alias name="unit" command="setup -- -q nosetests -s --with-id --nologcapture --tests=reahl/bzrsupport_dev -i '.*' -t"/>
