+++ reahl-dev/reahl_dev.egg-info/requires.txt
@@ -3,4 +3,3 @@ reahl-tofu>=4.0,<4.1
 Babel>=2.1,<2.1.999
 twine>=1.4.0,<1.4.9999
 wheel>=0.24.0,<0.24.9999
-reahl-bzrsupport>=4.0,<4.1
